pyproject.toml
[project]
name = "sonora-chat"
version = "1.0.0"
description = "Chat API for Sonora"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.34.0",
    "python-dotenv>=1.0.0",
    "python-multipart>=0.0.9",
    "openai>=1.50.0",
    "supabase>=2.10.0",
    "aiohttp>=3.10.0",
    "tiktoken>=0.8.0",
    "loguru>=0.7.0",
]

[dependency-groups]
dev = []
